chore: move url code to separate files
